maj uploadfichier

+++ Projet_quentin/projet_stage/UploadFichier.py
@@ -79,12 +79,16 @@ if __name__ == "__main__":
 				# Copier vers privé
 				private_path = os.path.join(private_dir, wav_file)
 				shutil.copy2(file_path, private_path)
+
+				# Supprimer l'original
+				os.remove(file_path)
 				
 			except Exception as e:
 				print(f"Erreur lors du traitement du fichier {wav_file}: {str(e)}")
 			continue
-		# Supprimer l'original
-		os.remove(file_path)
+
+		if wav_file.endswith('.wav'):
+			os.remove(file_path)
 		
 		# Pour les autres fichiers, procéder à l'upload dans le WordPress
 		result = upload_file_to_wordpress(file_path, WORDPRESS_URL, USERNAME, PASSWORD)
@@ -94,4 +98,3 @@ if __name__ == "__main__":
 			print(f"ID: {result.get('id')}")
 			print(f"URL: {result.get('source_url')}")
 			print(f"Titre: {result.get('title', {}).get('rendered', '')}")
-
